Split LogFormatter in RepositoryLogFormatter and PublicationLogFormatter, creating a BaseLogFormatter with common methods
